+++ js/dateUtils.js
@@ -1,0 +1,8 @@
+function monthsFromTo(d1, d2) {
+    var months;
+    months = (d2.getFullYear() - d1.getFullYear()) * 12;
+    months -= d1.getMonth();
+    months += d2.getMonth();
+    months++;
+    return months <= 0 ? 0 : months;
+}
